added comment to p152 with better solution; solution for p206 fixed some warnings ([[maybe_unused]] should not be needed in that case, but v142 thinks so) switched to v142 toolset from clang until the gtest nuget is fixed (or until switching to some other testing lib) fixed some mistakes in the solution configuration

+++ problems/p78.h
@@ -12,7 +12,7 @@ namespace p78
         {
             std::vector<std::vector<int>> pset;
 
-            size_t p = pow(2, nums.size());
+            size_t p = static_cast<size_t>(pow(2, nums.size()));
 
             for (size_t i = 0; i < p; ++i)
             {
